reshaping queries & loading database
