+++ tmc/queries/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["q_get_element_id",
+          "q_insert_into_tables",
+          "q_insert_adversary_x_tool",
+          "q_insert_tactic_x_technique",
+          "q_insert_tool_x_techn",
+          "q_insert_adversary_into_tables"
+           ]
